orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==2.9.0
